@@ -6,12 +6,12 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.2'
+        classpath 'com.android.tools.build:gradle:3.5.3'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
 
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
     }
 }
 
@@ -28,7 +28,7 @@ task clean(type: Delete) {
 }
 
 ext{
-    supportLibraryVersion = "27.0.0"
-    compileSdkVersion = 26
-    buildToolsVersion = '27.0.3'
+    supportLibraryVersion = "1.1.0"
+    compileSdkVersion = 29
+    buildToolsVersion = '29.0.0'
 }
