@@ -1,6 +1,6 @@
 package com.app.hubert.guide.lifecycle;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import com.app.hubert.guide.util.LogUtil;
 
